move pages into the app directory
